SettingsActivity is added

+++ SMediaLink/app/src/main/java/com/asuper/abocha/smedialink/Element.java
@@ -6,7 +6,7 @@ import android.widget.Button;
 public class Element {
 
     private int number;
-    private int green;
+    private float green;
 
     Element(int number, int green) {
         this.green = green;
@@ -21,11 +21,11 @@ public class Element {
         this.number = number;
     }
 
-    public int getGreen() {
+    public float getGreen() {
         return green;
     }
 
-    public void setGreen(int green) {
+    public void setGreen(float green) {
         this.green = green;
     }
 }
